requirements.txt
Django==4.0.3
djangorestframework==3.13.1
celery==5.2.3
whitenoise==2.0.6
gunicorn==20.1.0
django-celery-results==2.3.0
psycopg2-binary==2.9.3
